Cleanup account dialog, handle own nickname
I'd really like to have the nickname appear in the Login options, but I
was not able to figure out how to do that in libpurple, so right now
the Nickname will appear under advanced settings.

I'd also like to get a notification if someone changes the nickname
there and presses the save button - also, could not figure it out.

So for now this particular setting will only have an effect when logging
in.

Also added an account option so that changing nick names while being
logged on is possible.