Maybe show better info when DHT connection is lost
Well, neither purple_connection_update_progress() nor
purple_connection_notice() seem to show anything in Pidgin, it just sits
there pretending to be online.

We could show a network_connection_error, but that'd be too radical,
because libtox is actually trying to reconnect in the background and will
re-establish connection when possible.

So I guess not much more can be done here.

closes #6